Streamline flow: inline live cost, full scale range (1-200) + creativity
- Remove separate Scan & Estimate screen and confirmation popup; one "Upscale"
  button starts the job (validation/upload happen behind the scenes)
- Cost shown live next to the button, computed client-side from real dimensions
  ($0.016 x output MP), updating as scale/images change
- Scale factor is now the Crystal Upscaler's full 1-200 range (slider + number)
  instead of a 2x/4x toggle; add Creativity (0-10), both passed through to fal
- Backend: scale_factor/creativity as floats, clamped to model ranges; /api/config
  exposes ranges and $/MP; tests updated

+++ backend/app/settings/router.py
@@ -12,7 +12,7 @@ router = APIRouter(prefix="/api/settings", tags=["settings"])
 
 
 class SettingsUpdate(BaseModel):
-    default_scale_factor: int | None = None
+    default_scale_factor: float | None = None
     default_concurrency: int | None = None
     default_suffix: str | None = None
     default_output_format: str | None = None
